jwt/requirements.txt
flask
PyJWT
python-dotenv
psycopg2-binary
sqlalchemy